requirements.txt
aioredis==2.0.0
aiosqlite==0.17.0
alembic==1.7.1
appnope==0.1.2
asgiref==3.4.1
async-timeout==3.0.1
asyncpg==0.24.0
attrs==21.2.0
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
coverage==5.5
cypari==2.4.1
databases==0.5.2
decorator==5.0.9
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.3
fastapi==0.68.1
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
future==0.18.2
FXrays==1.3.5
greenlet==1.1.1
gunicorn==20.1.0
h11==0.12.0
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
iniconfig==1.1.1
ipython==7.27.0
iso8601==0.1.16
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
knot-floer-homology==1.1
Mako==1.1.5
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
networkx==2.6.3
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
plink==2.4
pluggy==1.0.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
ptyprocess==0.7.0
py==1.10.0
pydantic==1.8.2
Pygments==2.10.0
pyparsing==2.4.7
pypika-tortoise==0.1.1
pypng==0.0.21
pytest==6.2.5
pytest-asyncio==0.15.1
pytest-cov==2.12.1
python-dotenv==0.19.0
python-snappy==0.6.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
snappy-manifolds==1.1.2
spherogram==2.0
SQLAlchemy==1.4.23
starlette==0.14.2
toml==0.10.2
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.15.0
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.5.0
